+++ asdf/commands/diff.py
@@ -2,8 +2,12 @@
 Implementation of command for displaying differences between two ASDF files.
 """
 
+import argparse
 import sys
 from numpy import array_equal
+
+import jmespath
+
 try:
     # Provides cross-platform color support
     import colorama
@@ -44,9 +48,26 @@ class Diff(Command): # pragma: no cover
     """This class is the plugin implementation for the asdftool runner."""
     @classmethod
     def setup_arguments(cls, subparsers):
+        epilog = """
+examples:
+  diff two files:
+    asdftool diff file_before.asdf file_after.asdf
+  ignore differences in the file's ASDF metadata:
+    asdftool diff file_before.asdf file_after.asdf -i '[asdf_library,history]'
+  ignore differences in the 'foo' field of all objects in a list:
+    asdftool diff file_before.asdf file_after.asdf -i 'path.to.some_list[*].foo'
+
+See https://jmespath.org/ for more information on constructing
+JMESPath expressions.
+    """.strip()
+
         parser = subparsers.add_parser(
-            str("diff"), help="Report differences between two ASDF files",
-            description="""Reports differences between two ASDF files""")
+            "diff",
+            description="Report differences between two ASDF files",
+            epilog=epilog,
+            formatter_class=argparse.RawDescriptionHelpFormatter,
+            help="Report differences between two ASDF files",
+        )
 
         parser.add_argument(
             'filenames', metavar='asdf_file', nargs=2,
@@ -54,14 +75,18 @@ class Diff(Command): # pragma: no cover
 
         parser.add_argument(
             '-m', '--minimal', action='store_true',
-            help="Show minimal differences between the two files")
+            help="Show minimal differences between the two files.")
+
+        parser.add_argument(
+            '-i', '--ignore', action='append', dest='ignore',
+            help="JMESPath expression indicating tree nodes that should be ignored.")
 
         parser.set_defaults(func=cls.run)
         return parser
 
     @classmethod
     def run(cls, args):
-        return diff(args.filenames, args.minimal)
+        return diff(args.filenames, args.minimal, ignore=args.ignore)
 
 class ArrayNode:
     """This class is used to represent unique dummy nodes in the diff tree. In
@@ -110,12 +135,17 @@ class PrintTree:
 
 class DiffContext:
     """Class that contains context data of the diff to be computed"""
-    def __init__(self, asdf0, asdf1, iostream, minimal=False):
+    def __init__(self, asdf0, asdf1, iostream, minimal=False, ignore_ids=None):
         self.asdf0 = asdf0
         self.asdf1 = asdf1
         self.iostream = iostream
         self.minimal = minimal
         self.print_tree = PrintTree()
+
+        if ignore_ids is None:
+            self.ignore_ids = set()
+        else:
+            self.ignore_ids = ignore_ids
 
 def print_tree_context(diff_ctx, node_list, other, use_marker, last_was_list):
     """Print context information indicating location in ASDF tree."""
@@ -231,6 +261,9 @@ def compare_dicts(diff_ctx, dict0, dict1, keys, ignores=set()):
 
 def compare_trees(diff_ctx, tree0, tree1, keys=[]):
     """Recursively traverses two ASDF tree and compares them"""
+    if id(tree0) in diff_ctx.ignore_ids and id(tree1) in diff_ctx.ignore_ids:
+        return
+
     if both_are_ndarrays(tree0, tree1):
         compare_ndarrays(diff_ctx, tree0, tree1, keys)
     elif isinstance(tree0, dict) and isinstance(tree1, dict):
@@ -242,12 +275,48 @@ def compare_trees(diff_ctx, tree0, tree1, keys=[]):
     else:
         compare_objects(diff_ctx, tree0, tree1, keys)
 
-def diff(filenames, minimal, iostream=sys.stdout):
-    """Top-level implementation of diff algorithm"""
+def diff(filenames, minimal, iostream=sys.stdout, ignore=None):
+    """
+    Compare two ASDF files and write diff output to the stdout
+    or the specified I/O stream.
+
+    filenames : list of str
+        List of ASDF filenames to compare.  Must be length 2.
+
+    minimal : boolean
+        Set to True to forego some pretty-printing to minimize
+        the diff output.
+
+    iostream : io.TextIOBase, optional
+        Text-mode stream to write the diff, e.g., sys.stdout
+        or an io.StringIO instance.  Defaults to stdout.
+
+    ignore : list of str, optional
+        List of JMESPath expressions indicating tree nodes that
+        should be ignored.
+    """
+    if ignore is None:
+        ignore_expressions = []
+    else:
+        ignore_expressions = [jmespath.compile(e) for e in ignore]
+
     try:
         with asdf.open(filenames[0], _force_raw_types=True) as asdf0:
             with asdf.open(filenames[1], _force_raw_types=True) as asdf1:
-                diff_ctx = DiffContext(asdf0, asdf1, iostream, minimal=minimal)
+                ignore_ids = set()
+                for expression in ignore_expressions:
+                    for tree in [asdf0.tree, asdf1.tree]:
+                        result = expression.search(tree)
+                        if result is not None:
+                            ignore_ids.add(id(result))
+                        if isinstance(result, list):
+                            for elem in result:
+                                ignore_ids.add(id(elem))
+                        elif isinstance(result, dict):
+                            for value in result.values():
+                                ignore_ids.add(id(value))
+
+                diff_ctx = DiffContext(asdf0, asdf1, iostream, minimal=minimal, ignore_ids=ignore_ids)
                 compare_trees(diff_ctx, asdf0.tree, asdf1.tree)
     except ValueError as error:
         raise RuntimeError(str(error))